Serialize and Deserialize chain state (#11)
We want to allow to store prompt state half compiled. This is helpful
when using tool calls where user needs to receive a response from the AI that contains a tool call, do some work and respond back with the result of that work. In this scenario we need to serialize current compilation state to restart the compilation from that point.